OrganizerBot-master (1)/OrganizerBot-master/requirements.txt
watchdog==2.1.9
Pillow==9.5.0
pystray>=0.19.0
torch==2.0.1
transformers==4.30.2
python-telegram-bot>=20.0
opencv-python==4.7.0.72
scikit-learn==1.2.2
numpy==1.24.3
joblib==1.2.0
customtkinter==5.2.0 